simple shell workflow
